+++ apps/react-app/src/main.tsx
@@ -10,8 +10,8 @@
 
 async function enableMocking() {
   if (import.meta.env.DEV && import.meta.env.VITE_ENABLE_MSW === 'true') {
-    const { worker } = await import('@repo/mock/browser')
-    await worker.start({ onUnhandledRequest: 'bypass' })
+    const { startWorker } = await import('@repo/mock/browser')
+    await startWorker({ onUnhandledRequest: 'bypass' })
   }
 }